+++ green-house/frontend/src/Components/Content/How.css
@@ -1,0 +1,29 @@
+h1{
+    background-color: #58D592;
+}
+
+.howContainer{
+    background-color: #58D592;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-evenly;
+    height: 10cm;
+}
+
+.stepOne, .stepTwo, .stepThree{
+    display: flex;
+    background-color: var(--tertiary-color);
+    height: 200px;
+    width: 200px;
+    border-radius: 10%;
+    margin: 5px;
+    padding: 30px;
+    text-align: center;
+    font-weight: 500;
+}
+
+.howContainer p {
+  
+
+}
