+++ green-house/frontend/src/Components/Content/How.jsx
