+++ green-house/frontend/src/Components/Content/Why.css
@@ -7,8 +7,6 @@ h1{
     height: 10cm;
 }
 
-
-
 .whyContainer p {
     
     width: auto;
