--- green-house/frontend/src/Components/How/How.css
@@ -1,47 +1,0 @@
-h1{
-    background-color: #58D592;
-}
-
-.howContainer{
-    background-color: #58D592;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: space-evenly;
-    height: 10cm;
-}
-
-.stepOne{
-    display: flex;
-    background-color: var(--tertiary-color);
-    height: 200px;
-    width: 130px;
-    border-radius: 10%;
-    margin: 5px;
-    padding: 30px;
-}
-
-.stepTwo{
-    display: flex;
-    background-color: var(--tertiary-color);
-    height: 200px;
-    width: 130px;
-    border-radius: 10%;
-    margin: 5px;
-    padding: 30px;
-}
-
-.stepThree{
-    display: flex;
-    background-color: var(--tertiary-color);
-    height: 200px;
-    width: 130px;
-    border-radius: 10%;
-    margin: 5px;
-    padding: 30px;
-}
-
-.howContainer p {
-    text-align: left;
-
-}
